+++ apps/block_scout_web/lib/block_scout_web/views/api/rpc/eth_rpc_view.ex
@@ -26,7 +26,7 @@ defmodule BlockScoutWeb.API.RPC.EthRPCView do
 
     def encode(%BlockScoutWeb.API.RPC.EthRPCView{id: id, error: error}, _options) do
       """
-      {"jsonrpc":"2.0","error": #{error},"id": #{id}}
+      {"jsonrpc":"2.0","error": "#{error}","id": #{id}}
       """
     end
   end